refactor: fetchPosts 분리
